requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.7
django-redis==4.12.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1
